# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: DevDash API
  version: '2.7.0'
  description: |
    Self-hosted development control room. Tracks projects across machines, and —
    through an agent installed on each machine — provides remote terminals,
    filesystem access, process control, and chat-form control of AI coding CLIs.

    ## Authentication

    Every endpoint below requires a session except where noted. There are two
    equivalent credentials, both carrying the same JWT:

    - **Cookie** (`devdash-session`) — what the web UI uses. HttpOnly.
    - **Bearer token** — what native clients use.

    To obtain a Bearer token, send `X-DevDash-Client: native` on the
    `/auth` request. Without that header the token is issued **only** as a
    cookie: browsers deliberately never receive it in a readable place.

    Tokens last 30 days. After a week of age any authenticated `/api` request
    returns a fresh JWT in the `X-DevDash-Session` response header (and
    refreshes the cookie). Native clients must persist that header when it
    appears. On a 401, log in again.

    ## Login flow

    Login is a two-or-three step exchange, not one call:

    1. `POST /auth` `{action: "login", email, otpCode}`
       → `{pending: "2fa"}` (challenge issued), or
       → `{pending: "enroll"}` (2FA setup required), or
       → `{success: true, token}` (device already trusted)
    2. `POST /auth` `{action: "verify-2fa", code}` → `{success: true, token}`

...

    ## Realtime

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: devdash-session
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
    Machine:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        hidden: { type: boolean }
        terminalNameTemplate: { type: string, nullable: true }
        terminalPreviewLines: { type: integer }
      required: [id, name]
...
  responses:
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: >
        Authenticated, but not permitted — usually a machine you do not own,
        or 2FA enrollment not completed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    MachineOffline:
      description: The machine's agent is not connected, so it cannot be asked.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /api/machines:
    get:
      summary: List the user's machines
      responses:
        '200':
          description: Machines.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Machine' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/projects/{id}/process:
    post:
      summary: Start, stop or restart a project
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action: { type: string, enum: [start, stop, restart] }
              required: [action]
      responses:
        '200': { description: Command dispatched to the agent. }
        '503': { $ref: '#/components/responses/MachineOffline' }
...
  /api/browse:
    post:
      summary: List a directory on a machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                machineId: { type: integer }
                path: { type: string }
      responses:
        '200':
          description: Directory entries.
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/MachineOffline' }
...